Refactoring PES Class
