+++ CMakeLists.txt
@@ -9,13 +9,13 @@ add_executable(LRF_Integration
                     Cpp/PotentialEnergySurface.h
                     Cpp/PotentialEnergySurface.cpp
                     Cpp/interactions.cpp
-                    Cpp/evaluate_lrf.cpp
-                    Cpp/calculate_tensor.cpp
-        )
+        Cpp/tensor.cpp
+        Cpp/tensor.h
+)
 target_link_libraries(LRF_Integration)
 
 
 #How to run it
 #cmake -S . -B build
 #cmake --build build -j 12
-# for testing only cd build && ctest --output-on-failure -j12
+# for testing only: cd build && ctest --output-on-failure -j12
